% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateCells.R
\name{AnnotateCells}
\alias{AnnotateCells}
\title{Extract Spatial Ecotype Annotations for Single Cells}
\usage{
AnnotateCells(scmeta, obj, col = "SE", dropcell = TRUE)
}
\arguments{
\item{scmeta}{Data frame containing single-cell metadata. Two columns (X and Y) for spatial coordinates are required.}

\item{obj}{A Seurat object returned from SpatialEcoTyper function.}

\item{col}{Character string specifying the name of the column in `obj@meta.data`
containing the spatial ecotype annotations. Default is `"SE"`.}

\item{dropcell}{Logical. If TRUE, cells that cannot be assigned to any spatial
ecotype (outside the radius) will be removed from the returned metadata. Default is TRUE.}
}
\value{
An updated version of `scmeta` with spatial ecotype annotations added.
}
\description{
This function adds spatial ecotype annotations to the metadata of single cells.
}
\examples{
library(data.table)
library(Seurat)
library(SpatialEcoTyper)

scmeta <- read.table("https://spatialecotyper.stanford.edu/inc/inc.public.vignettes.php?file=Melanoma1_subset_scmeta.tsv",
                     sep = "\t", header = TRUE, row.names = 1)
head(scmeta)

url <- "https://spatialecotyper.stanford.edu/inc/inc.public.vignettes.php?file=Melanoma1_subset_SpatialEcoTyper_results.rds"
download.file(url, destfile = "Melanoma1_subset_SpatialEcoTyper_results.rds", mode = "wb")
obj <- readRDS("Melanoma1_subset_SpatialEcoTyper_results.rds")$obj

## Transfer SE annotations to single cells
scmeta <- AnnotateCells(scmeta = scmeta, obj = obj, dropcell = TRUE)
head(scmeta)

}
