% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeAvgs.R
\name{ComputeAvgs}
\alias{ComputeAvgs}
\title{Compute Cell-Type-Specific Average Expression of Spatial Clusters}
\usage{
ComputeAvgs(
  normdata,
  scmeta,
  cluster = "SE",
  Region = NULL,
  scale = TRUE,
  ncores = 4
)
}
\arguments{
\item{normdata}{Numeric matrix of normalized expression data, where rows
represent genes and columns represent cells.}

\item{scmeta}{Data frame containing metadata associated with each cell,
including spatial cluster and cell type annotations.}

\item{cluster}{Character string specifying the column name in 'scmeta'
containing spatial cluster annotations.}

\item{Region}{Character string specifying the column name in metadata data
frames containing region annotations (default: NULL).}

\item{scale}{A boolean specifying whether to do univariance normalization.}

\item{ncores}{Integer specifying the number of CPU cores to use for parallel processing.}
}
\value{
A matrix of average expression, where rows represent genes and columns
represent spatial clusters from the sample.
}
\description{
Compute Cell-Type-Specific Average Expression of Spatial Clusters
}
\examples{
library(data.table)
library(Seurat)
library(SpatialEcoTyper)

scdata <- fread("https://spatialecotyper.stanford.edu/inc/inc.public.vignettes.php?file=Melanoma1_subset_counts.tsv.gz",
                sep = "\t",header = TRUE, data.table = FALSE)
rownames(scdata) <- scdata[, 1]
scdata <- as.matrix(scdata[, -1])
tmpobj <- CreateSeuratObject(scdata) \%>\% NormalizeData()
normdata <- GetAssayData(tmpobj, layer = "data")

url <- "https://spatialecotyper.stanford.edu/inc/inc.public.vignettes.php?file=Melanoma1_subset_SpatialEcoTyper_results.rds"
download.file(url, destfile = "Melanoma1_subset_SpatialEcoTyper_results.rds", mode = "wb")
metadata = readRDS("Melanoma1_subset_SpatialEcoTyper_results.rds")$metadata

# Construct cell-type-specific gene expression signatures of SEs
avgexprs <- ComputeAvgs(normdata = normdata, scmeta = metadata)
head(avgexprs)

}
\keyword{internal}
