% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatePseudobulks.R
\name{CreatePseudobulks}
\alias{CreatePseudobulks}
\title{Create Pseudo-bulk Mixtures}
\usage{
CreatePseudobulks(data = NULL, groups, counts = NULL, n_mixtures = 100)
}
\arguments{
\item{data}{A matrix of normalized gene expression data (genes x cells). If NULL, counts must be provided.}

\item{groups}{A named vector indicating the group (e.g., spatial ecotype) for each cell.
The names should correspond to the column names of the data matrix.}

\item{counts}{A matrix of raw counts data (genes x cells). Used to generate normalized data if data is not provided.}

\item{n_mixtures}{An integer specifying the number of pseudobulk samples to create. Default is 100.}
}
\value{
A list containing two elements:
\item{Fracs}{A matrix of the fractions of each group in the pseudobulk samples (rows represent pseudobulk samples, columns represent groups).}
\item{Mixtures}{A matrix of pseudobulk gene expression data (genes x pseudobulk samples).}
}
\description{
This function generates pseudobulk samples by aggregating single-cell transcriptomics.
}
\details{
\itemize{
  \item If `data` is not provided, the function will normalize the `counts` matrix by dividing each column by its sum and multiplying by 10,000.
  \item If the maximum value in `data` is less than 80, it assumes the data is in log2 scale and converts it back to non-log scale.
  \item The `groups` vector is used to ensure that cells are correctly assigned to their respective groups.
  If `groups` does not have names, it is assumed that the names correspond to the column names of the `data` matrix.
  \item Pseudobulk samples are created by sampling cells from each group based on predefined fractions,
  and then calculating the average expression for each gene in the pseudobulk samples.
  \item The pseudobulk data is then normalized using Seurat's `NormalizeData` function.
}
}
\examples{
library(SpatialEcoTyper)
library(googledrive)
drive_deauth() # no Google sign-in is required
drive_download(as_id("15n9zlXed74oeGaO1pythOOM_iWIfuMn2"), "Melanoma_WU2161_counts.rds",
                    overwrite = TRUE)
counts <- readRDS("Melanoma_WU2161_counts.rds") ## raw counts of scRNA-seq data
groups <- sample(paste0("SE", 1:10), ncol(counts), replace = TRUE)
names(groups) <- colnames(counts)
result = CreatePseudobulks(counts = counts, groups = groups, n_mixtures = 20)
head(result$Mixtures[, 1:5]) ## Gene expression matrix of pseudobulks
head(result$Fracs) ## SE fractions in pseudobulks

}
