% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeconvoluteSE.R
\name{DeconvoluteSE}
\alias{DeconvoluteSE}
\title{Infer SE Abundances Using a Pretrained NMF Model}
\usage{
DeconvoluteSE(
  dat,
  scale = TRUE,
  W = NULL,
  nsample.per.run = 500,
  sum2one = TRUE,
  ncores = 1
)
}
\arguments{
\item{dat}{A numeric matrix of gene expression data, e.g. TPM matrix for bulk RNA-seq data, logCPM matrix for Visium data etc.}

\item{scale}{Logical. If `TRUE`, the input data is scaled before making predictions.}

\item{W}{A numeric matrix representing the basis matrix (`W`) from a pretrained NMF model.}

\item{nsample.per.run}{Integer specifying the maximum number of samples to process per
prediction run to manage memory usage efficiently.}

\item{sum2one}{Logical. If `TRUE`, normalizes the predicted SE abundances so that
they sum to 1 for each sample.}

\item{ncores}{Integer specifying the number of CPU cores to use for parallel processing.}
}
\value{
A matrix of SE abundances in bulk tumors.
}
\description{
This function estimates the abundances of SEs in bulk tumor gene expression data
using a provided basis matrix (`W`) from NMF model.
}
\examples{
library(googledrive)
drive_deauth() # no Google sign-in is required
drive_download(as_id("14QvmgISxaArTzWt_UHvf55aAYN2zm84Q"), "SKCM_RNASeqV2.geneExp.rds",
                    overwrite = TRUE)
bulkdata <- readRDS("SKCM_RNASeqV2.geneExp.rds")

# Predict SE abundances in bulk tumors
se_abundances <- DeconvoluteSE(dat = bulkdata)
head(se_abundances[, 1:5])

}
