% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSNList.R
\name{getSN}
\alias{getSN}
\title{Construct Similarity Network}
\usage{
getSN(emb, npcs = 20, k = 50)
}
\arguments{
\item{emb}{A matrix contains the embedding, with rows correspond to principal
components (PCs) or features, and columns correspond to spatial neighborhoods.}

\item{npcs}{An integer specifying the number of principal components or features
for the analysis. Default is 20.}

\item{k}{An integer specifying the number of nearest neighbors to consider when constructing the KNN graph.
Default is 50.}
}
\value{
A matrix for the similarity network (SN). The rows and columns represent spatial neighborhood,
and the values represent inverted distance.
}
\description{
This function generates a similarity network (SN) for a given embeddings.
Specifically, the function computes K-nearest neighbor (KNN) graphs based
on the provided embedding.
}
\details{
The function computes a KNN graph using the input embedding and inverts the resulting
distances to similarity scores.
}
