% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferNCells.R
\name{InferNCells}
\alias{InferNCells}
\title{Infer Cell Numbers per Spot from Expression Profiles}
\usage{
InferNCells(normdat, avg.number = 5)
}
\arguments{
\item{normdat}{A numeric matrix or data frame with genes in rows and
spots in columns. Values should represent normalized or raw
expression counts.}

\item{avg.number}{Numeric value specifying the desired average number
of inferred cells per spot. Default is 5 for Visium.}
}
\value{
An integer vector containing the inferred number of cells for
  each spot
}
\description{
Estimates the number of cells for each Visium spot based on
total gene expression. The inferred values are linearly scaled such
that the spot with the lowest total expression is assigned 1 cell,
and the average inferred number of cells across all spots is approximately
equal to `avg.number`.
}
