% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntegrateSpatialEcoTyper.R
\name{IntegrateSpatialEcoTyper}
\alias{IntegrateSpatialEcoTyper}
\title{Integrate Multiple Spatial Transcriptomics Datasets to Identify Conserved Spatial Ecotypes}
\usage{
IntegrateSpatialEcoTyper(
  SpatialEcoTyper_list,
  data_list,
  outdir = "./",
  normalization.method = "None",
  nmf_ranks = 10,
  nrun.per.rank = 30,
  min.coph = 0.95,
  nfeatures = 3000,
  min.features = 10,
  Region = NULL,
  downsample.by.region = TRUE,
  subresolution = 30,
  minibatch = 5000,
  ncores = 1,
  seed = 1
)
}
\arguments{
\item{SpatialEcoTyper_list}{A named list of SpatialEcoTyper results, each item
represents a list returned from the \code{SpatialEcoTyper} function.}

\item{data_list}{A named list of expression matrices where each matrix represents
gene expression data used for the SpatialEcoTyper analysis. The list name should match
that of \code{SpatialEcoTyper_list}.}

\item{outdir}{Directory where the results will be saved. Defaults to the current
directory with a subdirectory named "SpatialEcoTyper_results_" followed by the current date.}

\item{normalization.method}{Method for normalizing the expression data. Options
include "None" (default), "SCT", or other methods compatible with Seurat's `NormalizeData` function.}

\item{nmf_ranks}{Integer or a vector specifying the number of clusters (10 by default).
When an integer vector is supplied, the function will test all supplied numbers and
select the optimal number, which takes time.}

\item{nrun.per.rank}{An integer specifying the the number of runs per rank for NMF (default: 30).}

\item{min.coph}{Numeric specifying the minimum cophenetic coefficient required for a rank to be optimal.}

\item{nfeatures}{An integer specifying the maximum number of top variable genes to select for each cell type.}

\item{min.features}{An integer specifying the minimum number of shared features (genes) required across samples.}

\item{Region}{A character string specifying the column name in metadata data frames
containing region annotations (default: NULL). Pathologist annotation is recommended if available.}

\item{downsample.by.region}{A Boolean indicating whether to perform downsampling
to balance the number of spatial clusters across regions.}

\item{subresolution}{A numeric specifying the resolution for clustering within each sample.}

\item{minibatch}{Integer specifying the number of columns to process in each minibatch in the SNF analysis.
Default is 5000. This option splits the matrix into smaller chunks (minibatch), thus reducing memory usage.}

\item{ncores}{An integer specifying the number of cores for parallel processing. Default is 1.}

\item{seed}{An integer used to seed the random number generator.}
}
\description{
Integrate Multiple Spatial Transcriptomics Datasets to Identify Conserved Spatial Ecotypes
}
\examples{
# See https://digitalcytometry.github.io/spatialecotyper/docs/articles/Integration.html

}
