% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mostFrequent}
\alias{mostFrequent}
\title{Identify the most frequent category in a vector}
\usage{
mostFrequent(x)
}
\arguments{
\item{x}{A vector of values (e.g., categorical or character data).}
}
\value{
A character string representing the most frequent value in the input vector.
}
\description{
This function takes a vector as input and returns the most frequent
category or value. It works by converting the input into a table of
frequencies, sorting the frequencies in descending order, and returning
the most common value as a character string.
}
