% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFGenerateW.R
\name{NMFGenerateW}
\alias{NMFGenerateW}
\title{Train SE Deconvolution Model}
\usage{
NMFGenerateW(
  Fracs,
  ExpMat,
  scale = TRUE,
  nfeature = 2000,
  nfeature.per.se = 50,
  method = "brunet"
)
}
\arguments{
\item{Fracs}{A fraction matrix, with rows as samples.}

\item{ExpMat}{A gene expression matrix with genes in rows and samples in columns.}

\item{scale}{Logical indicating whether to scale the gene expression matrix. Default is TRUE.}

\item{nfeature}{Integer specifying the top variable features for training the models (default: 2000).}

\item{nfeature.per.se}{Integer specifying the maximal number of features to select for each SE (default: 50).}

\item{method}{A character string specifying the NMF method to use. Default is "brunet".}
}
\value{
A matrix containing the basis NMF W matrix with rows as features and columns as SEs.
}
\description{
This function trains a Non-negative Matrix Factorization (NMF) model for SE deconvolution
based on given spatial ecotype fractions and gene expression matrix. Prior to NMF, each gene
is scaled to mean 0 and unit variance. To satisfy the non-negativity requirement of NMF,
the expression matrix is processed using posneg transformation, which converts the expression
matrix into two matrices, one containing only positive values and the other containing only
negative values with the sign inverted. The two matrices are subsequently concatenated to
produce the training data.
}
