% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFpredict.R
\name{NMFpredict}
\alias{NMFpredict}
\title{Prediction Using Pretrained NMF Model}
\usage{
NMFpredict(
  W,
  testdat,
  scale = FALSE,
  ncell.per.run = 500,
  sum2one = TRUE,
  ncores = 1
)
}
\arguments{
\item{W}{Matrix representing the factorization matrix W of a pretrained NMF model.}

\item{testdat}{Numeric matrix containing the new data for which NMF scores are to be predicted.}

\item{scale}{Logical indicating whether to scale the input data.}

\item{ncell.per.run}{Integer specifying the maximum number of cells per NMF prediction run to avoid memory issues.}

\item{sum2one}{Logical. If `TRUE`, normalizes the predicted scores so that
they sum to 1 for each sample.}

\item{ncores}{Integer specifying the number of CPU cores to use for parallel processing.}
}
\value{
A matrix representing the NMF prediction scores, with rows as samples/cells and columns as SE groups.
}
\description{
This function uses pretrained NMF models to recover cell states / spatial ecotypes.
It takes a factorization matrix W representing a pretrained NMF model and a numeric gene expression matrix.
}
\examples{

bulkdata <- fread("https://spatialecotyper.stanford.edu/inc/inc.public.vignettes.php?file=SKCM_RNASeqV2.geneExp.tsv",
                  sep = "\t", header = TRUE, data.table = FALSE)
rownames(bulkdata) = bulkdata[, 1]
bulkdata = as.matrix(bulkdata[, -1])
W <- readRDS(file.path(system.file("extdata", package = "SpatialEcoTyper"), "Bulk_SE_Recovery_W.rds"))

# Predict SE abundances in bulk tumors
preds <- NMFpredict(W = W, bulkdata, scale = TRUE)
head(preds[, 1:5])

}
