% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitionTissue.R
\name{PartitionTissue}
\alias{PartitionTissue}
\title{Partition Tissue into Subregions}
\usage{
PartitionTissue(meta, nrow = 2, ncol = 2, X = "X", Y = "Y")
}
\arguments{
\item{meta}{A data frame containing metadata for the tissue samples. It must include
columns representing spatial coordinates (`X` and `Y`).}

\item{nrow}{An integer specifying the number of rows to divide the tissue into. Default is 2.}

\item{ncol}{An integer specifying the number of columns to divide the tissue into. Default is 2.}

\item{X}{A string representing the name of the column containing X coordinates. Default is "X".}

\item{Y}{A string representing the name of the column containing Y coordinates. Default is "Y".}
}
\value{
A modified data frame with an additional column, `Partition`, that labels each row
        with the corresponding subregion identifier in the format "row_column".
}
\description{
This function partitions a tissue dataset into smaller subregions based on spatial coordinates.
The partitioning is determined by dividing the range of the X and Y coordinates into
specified numbers of rows (`nrow`) and columns (`ncol`). Each subregion is assigned a
unique identifier, facilitating independent analysis of smaller subsets of the tissue.
}
\examples{
# Example metadata with spatial coordinates
meta <- data.frame(
  X = c(10, 20, 30, 40, 50),
  Y = c(15, 25, 35, 45, 55)
)

# Partition the tissue into 2 rows and 2 columns
PartitionTissue(meta, nrow = 2, ncol = 2)

}
\keyword{internal}
