% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessST.R
\name{PreprocessST}
\alias{PreprocessST}
\title{Preprocess Spatial Transcriptomics Data}
\usage{
PreprocessST(
  expdat,
  metadata,
  min.cells = 3,
  min.features = 5,
  X = "X",
  Y = "Y"
)
}
\arguments{
\item{expdat}{A matrix or data frame representing the gene expression data,
where rows correspond to genes and columns correspond to cells.}

\item{metadata}{A data frame containing metadata associated with each cell.
Must include spatial coordinates (e.g., X and Y) as well as other cell-specific
annotations. The row names of the `metadata` must match the column names of the `expdat`.}

\item{min.cells}{An integer specifying the minimum number of cells in which a
gene must be expressed to be retained (default is 3).}

\item{min.features}{An integer specifying the minimum number of features (genes)
a cell must express to be retained (default is 5).}

\item{X}{A string specifying the column name in the metadata data frame that
represents the X spatial coordinate (default is "X").}

\item{Y}{A string specifying the column name in the metadata data frame that
represents the Y spatial coordinate (default is "Y").}
}
\value{
A list containing two elements:
\describe{
  \item{expdat}{A filtered matrix of gene expression data, converted to a sparse matrix}
  \item{metadata}{A filtered data frame of metadata, aligned with the filtered gene expression data, including reformatted spatial coordinates.}
}
}
\description{
This function preprocesses single-cell spatial transcriptomics data by filtering
out low-quality genes and cells based on specified thresholds. It ensures that
only genes expressed in a minimum number of cells and cells expressing a minimum
number of features are retained. Additionally, it reformats the metadata to include
spatial coordinates (X and Y).
}
\examples{
library(SpatialEcoTyper)
library(data.table)

scdata <- fread("https://spatialecotyper.stanford.edu/inc/inc.public.vignettes.php?file=Melanoma1_subset_counts.tsv.gz",
                sep = "\t",header = TRUE, data.table = FALSE)
rownames(scdata) <- scdata[, 1]
scdata <- as.matrix(scdata[, -1])
scmeta <- read.table("https://spatialecotyper.stanford.edu/inc/inc.public.vignettes.php?file=Melanoma1_subset_scmeta.tsv",
                     sep = "\t", header = TRUE, row.names = 1)
processed <- PreprocessST(expdat = scdata, scmeta, X = "X", Y = "Y",
                          min.cells = 3, min.features = 5)
head(processed$metadata)
head(processed$expdat)

}
