% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{rankSparse}
\alias{rankSparse}
\title{Transform a Sparse Matrix to Rank Space (Rank Non-zeros in Each Column)}
\usage{
rankSparse(sparseMat)
}
\arguments{
\item{sparseMat}{A sparse matrix}
}
\value{
A sparse matrix of the same dimensions as the input, where the non-zero elements
        are replaced by their ranks within each column, normalized by the number of rows.
}
\description{
This function ranks the non-zero elements of a sparse matrix within each column.
The ranks are scaled by dividing them by the number of rows in the matrix,
resulting in ranks normalized between 0 and 1. The original zero values are left unchanged.
}
\examples{
library(Matrix)
# Create a sample sparse matrix
sparseMat <- rsparsematrix(5, 5, density = 0.5)
# Apply the ranking function
rankedSparseMat <- rankSparse(sparseMat)
print(rankedSparseMat)

}
