% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNF2.R
\name{SNF2}
\alias{SNF2}
\title{Enhanced Similarity Network Fusion}
\usage{
SNF2(Wall, K = 10, t = 10, minibatch = 5000, ncores = 4, verbose = FALSE)
}
\arguments{
\item{Wall}{List of similarity matrices. Each element of the list is a square,
symmetric matrix that shows affinities of the data points from a certain view.}

\item{K}{Number of neighbors in K-nearest neighbors part of the algorithm.}

\item{t}{Number of iterations for the diffusion process.}

\item{minibatch}{Integer specifying the number of columns to process in each minibatch.
Default is 5000. This option splits the matrix into smaller chunks (minibatch),
thus reducing memory usage.}

\item{ncores}{Integer specifying the number of CPU cores to use for parallel processing.}

\item{verbose}{Boolean specifying whether to show the progress messages.}
}
\value{
A fused matrix.
}
\description{
Similarity Network Fusion (SNF) integrates multiple views (similarity matrices) to
construct an overall status matrix. This function is adopted from the SNFtool
(https://github.com/maxconway/SNFtool/) and has been enhanced for unsupervised analysis of spatial ecosystem.
The new function supports sparse matrix and missing data in the input matrices.
}
