% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialView.R
\name{SpatialView}
\alias{SpatialView}
\title{Visualize Spatial Landscape of Cells / Spots}
\usage{
SpatialView(
  obj,
  by,
  X = "X",
  Y = "Y",
  pt.shape = 20,
  pt.size = 0.5,
  pt.alpha = 1,
  jitter = FALSE,
  slot = "data",
  coord.fix = FALSE,
  highlight.cells = NULL,
  control.cells = NULL,
  ncol = 3,
  bg.downsample = 2000,
  bg.color = "gray80",
  bg.size = 0.5,
  bg.alpha = 0.7
)
}
\arguments{
\item{obj}{A Seurat object or a data frame (with cell names as row names).}

\item{by}{A feature name for plotting, e.g. cell type, region, gene expression.}

\item{X}{A character specifying the spatial coordinate of x-axis.}

\item{Y}{A character specifying the the spatial coordinate of y-axis.}

\item{pt.shape}{Point shape for plotting}

\item{pt.size}{A numeric specifying the point size of non-control cells.}

\item{pt.alpha}{A numeric specifying the point transparency of non-control cells.}

\item{jitter}{A boolean specifying whether add jitters to the cells.}

\item{slot}{The slot in Seurat object to pull feature from.}

\item{highlight.cells}{A vector specifying the cells for highlighting.}

\item{control.cells}{A vector specifying the control cells as background. If not specified,
all the non-highlighting cells will be considered as control.cells.}

\item{bg.downsample}{An integer specifying the aim for downsampling the control.cells.}

\item{bg.color}{Color of control cells.}

\item{bg.size}{A numeric specifying the point size of control cells.}

\item{bg.alpha}{A numeric specifying the point transparency of control cells.}
}
\value{
A ggplot object.
}
\description{
Visualize Spatial Landscape of Cells / Spots
}
\examples{
library(data.table)
library(Seurat)
library(SpatialEcoTyper)
library(ggplot2)
library(googledrive)
drive_deauth() # no Google sign-in is required
drive_download(as_id("1CgUOQKrWY_TG61o5aw7J9LZzE20D6NuI"),
                    "HumanMelanomaPatient1_subset_scmeta.tsv", overwrite = TRUE)
scmeta <- fread("HumanMelanomaPatient1_subset_scmeta.tsv",
                sep = "\t",header = TRUE, data.table = FALSE)

# Visualize the cell type annotations in the tissue
SpatialView(scmeta, by = "CellType", X = "X", Y = "Y") +
            scale_color_manual(values = pals::kelly()[-1])
SpatialView(scmeta, by = "Region", X = "X", Y = "Y") +
            scale_color_brewer(type = "qual", palette = "Set1")

}
