% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Znorm.R
\name{Znorm}
\alias{Znorm}
\title{Weighted / Unweighted Uni-variance Normalization}
\usage{
Znorm(mat, groups = NULL)
}
\arguments{
\item{mat}{A matrix of gene expression data, where rows represent genes and columns represent cells.}

\item{groups}{A character vector specifying the group labels for each cell.
If provided, weighted univariance normalization will be performed.}
}
\value{
A matrix of Z-score normalized gene expression data, with rows representing genes and columns representing cells.
}
\description{
Z-score normalization scales each feature (gene) across all cells to have a mean of 0 and a standard deviation of 1.
This function allows for optional weighted univariance normalization, where cells are grouped by a categorical variable.
}
\examples{
library(Seurat)
library(data.table)
library(SpatialEcoTyper)
scdata <- fread("https://spatialecotyper.stanford.edu/inc/inc.public.vignettes.php?file=Melanoma1_subset_counts.tsv.gz",
                sep = "\t",header = TRUE, data.table = FALSE)
rownames(scdata) <- scdata[, 1]
scdata <- as.matrix(scdata[, -1])
tmpobj <- CreateSeuratObject(scdata) \%>\% NormalizeData(verbose = FALSE)
normdata <- GetAssayData(tmpobj, layer = "data")
# Z-score normalization
znorm_data <- Znorm(normdata)
head(znorm_data[, 1:5])

# Weighted Z-score normalization
scmeta <- fread("https://spatialecotyper.stanford.edu/inc/inc.public.vignettes.php?file=Melanoma1_subset_scmeta.tsv",
                sep = "\t",header = TRUE, data.table = FALSE)
wtdznorm_data <- Znorm(normdata, groups = scmeta$Region)
head(wtdznorm_data[, 1:5])

}
