% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateCells.R
\name{AnnotateCells}
\alias{AnnotateCells}
\title{Extract Spatial Ecotype Annotations for Single Cells}
\usage{
AnnotateCells(scmeta, obj, col = "SE", dropcell = TRUE)
}
\arguments{
\item{scmeta}{Data frame containing single-cell metadata. Two columns (X and Y) for spatial coordinates are required.}

\item{obj}{A Seurat object returned from SpatialEcoTyper function.}

\item{col}{Character string specifying the name of the column in `obj@meta.data`
containing the spatial ecotype annotations. Default is `"SE"`.}

\item{dropcell}{Logical. If TRUE, cells that cannot be assigned to any spatial
ecotype (outside the radius) will be removed from the returned metadata. Default is TRUE.}
}
\value{
An updated version of `scmeta` with spatial ecotype annotations added.
}
\description{
This function adds spatial ecotype annotations to the metadata of single cells.
}
\examples{
library(data.table)
library(Seurat)
library(SpatialEcoTyper)
library(googledrive)
drive_deauth() # no Google sign-in is required
## Download a single-cell meta data
drive_download(as_id("1CgUOQKrWY_TG61o5aw7J9LZzE20D6NuI"),
                    "HumanMelanomaPatient1_subset_scmeta.tsv",
                     overwrite = TRUE)
## Obtain a Seurat object from SpatialEcoTyper result
drive_download(as_id("1nSPj2zRywFUdbo1fwiz77ds4NuM6bmV2"),
              "Melanoma1_subset_SpatialEcoTyper_results.rds", overwrite = TRUE)

scmeta <- read.table("HumanMelanomaPatient1_subset_scmeta.tsv",
                     sep = "\t", header = TRUE, row.names = 1)
head(scmeta)
obj <- readRDS("Melanoma1_subset_SpatialEcoTyper_results.rds")$obj

## Transfer SE annotations to single cells
scmeta <- AnnotateCells(scmeta = scmeta, obj = obj, dropcell = TRUE)
head(scmeta)

}
