% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatmapView.R
\name{drawRectangleAnnotation}
\alias{drawRectangleAnnotation}
\title{Draw Rectangle Annotations}
\usage{
drawRectangleAnnotation(rows, columns, col = "black")
}
\arguments{
\item{rows}{A vector of row identifiers.}

\item{columns}{A vector of column identifiers.}
}
\description{
This function draw rectangle grids based on given row and column factors.
}
\examples{
library(grid)
library(SpatialEcoTyper)
library(ComplexHeatmap)

dat = matrix(rnorm(100), 10)
rownames(dat) = letters[1:10]
colnames(dat) = letters[11:20]
rowann = data.frame(Group = rep(letters[1:2], each=5), index = 1:10)
colann = data.frame(Group = rep(letters[1:2], each=5), index = 11:20)
HeatmapView(dat, left_ann = rowann, top_ann = colann)
drawRectangleAnnotation(rowann$Group, colann$Group)

}
