% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSpatialMetacells.R
\name{GetSpatialMetacells}
\alias{GetSpatialMetacells}
\title{Construct Spatial Metacells from Single-Cell Spatial Data}
\usage{
GetSpatialMetacells(
  normdata,
  metadata,
  X = "X",
  Y = "Y",
  CellType = "CellType",
  spotCoord = NULL,
  k = 20,
  radius = 50,
  min.cells.per.region = 1,
  ncores = 4
)
}
\arguments{
\item{normdata}{A matrix or data frame representing normalized gene expression data,
where rows correspond to genes and columns correspond to cells.}

\item{metadata}{A data frame containing metadata associated with each cell.
Must include spatial coordinates (e.g., X and Y) as well as cell type
annotations. The row names of the \code{metadata} must match the column names of the \code{normdata}.}

\item{X}{Character string specifying the column name in \code{metadata} containing the X spatial coordinates.}

\item{Y}{Character string specifying the column name in \code{metadata} containing the Y spatial coordinates.}

\item{CellType}{Character string specifying the column name in \code{metadata} containing the cell type annotations.}

\item{spotCoord}{A data frame containing the spatial coordinates (\code{X} and \code{Y}) of each spatial neighborhood,
with neighborhood identifiers as row names.}

\item{k}{Integer specifying the number of nearest spatial neighbors for constructing metacells.}

\item{radius}{Numeric value specifying the radius (in units of spatial coordinates) within which neighboring cells are considered.}

\item{min.cells.per.region}{Integer specifying the minimum number of cells required in each spatial neighborhood to compute metacells.}

\item{ncores}{Integer specifying the number of CPU cores to use for parallel processing (default: 4).}
}
\value{
A matrix containing the spatial metacell expression profiles, with rows representing individual genes
        and columns representing metacells.
}
\description{
This function computes spatial metacells from single-cell spatial transcriptomics.
Spatial metacells represent aggregated expression profiles for spatially proximal cells.
The function utilizes k-nearest neighbor (KNN) weighting to aggregate expression profiles from neighboring cells
and generate metacell expression profiles for each cell type.
}
\examples{
library(data.table)
library(Seurat)
library(SpatialEcoTyper)
scdata <- fread("https://spatialecotyper.stanford.edu/inc/inc.public.vignettes.php?file=Melanoma1_subset_counts.tsv.gz",
                sep = "\t",header = TRUE, data.table = FALSE)
rownames(scdata) <- scdata[, 1]
scdata <- as.matrix(scdata[, -1])
normdata = NormalizeData(scdata)
scmeta <- read.table("https://spatialecotyper.stanford.edu/inc/inc.public.vignettes.php?file=Melanoma1_subset_scmeta.tsv",
                      sep = "\t",header = TRUE, row.names = 1)

# Construct spatial metacells around B cells
snmeta = scmeta[scmeta$CellType=="B", c("X", "Y")]
metacells <- GetSpatialMetacells(normdata = normdata,
                                 metadata = scmeta,
                                 X = "X", Y = "Y",
                                 CellType = "CellType",
                                 spotCoord = snmeta)
head(metacells)

}
