% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSpatialMetacells.R
\name{GetSpatialMetacells}
\alias{GetSpatialMetacells}
\title{Construct Spatial Metacells from Single-Cell Spatial Data}
\usage{
GetSpatialMetacells(
  normdata,
  metadata,
  X = "X",
  Y = "Y",
  CellType = "CellType",
  k = 20,
  radius = 50,
  bin = "SpotID",
  bin.X = "Spot.X",
  bin.Y = "Spot.Y",
  min.cells.per.region = 1,
  ncores = 1
)
}
\arguments{
\item{normdata}{A matrix or data frame representing normalized gene expression data,
where rows correspond to genes and columns correspond to cells.}

\item{metadata}{A data frame containing metadata associated with each cell.
Must include spatial coordinates (e.g., X and Y) as well as cell type
annotations. The row names of the \code{metadata} must match the column names of the \code{normdata}.}

\item{X}{Character string specifying the column name in \code{metadata} containing the X spatial coordinates.}

\item{Y}{Character string specifying the column name in \code{metadata} containing the Y spatial coordinates.}

\item{CellType}{Character string specifying the column name in \code{metadata} containing the cell type annotations.}

\item{k}{Integer specifying the number of nearest spatial neighbors for constructing metacells.}

\item{radius}{Numeric value specifying the radius (in units of spatial coordinates) within which neighboring cells are considered.}

\item{bin}{Character string specifying the column name in \code{metadata} containing the spatial neighborhood identifier.}

\item{bin.X}{Character string specifying the column name in \code{metadata} containing the X spatial coordinates of spatial neighborhoods.}

\item{bin.Y}{Character string specifying the column name in \code{metadata} containing the Y spatial coordinates of spatial neighborhoods.}

\item{min.cells.per.region}{Integer specifying the minimum number of cells required in each spatial neighborhood to compute metacells.}

\item{ncores}{Integer specifying the number of CPU cores to use for parallel processing (default: 1).}
}
\value{
A matrix containing the spatial metacell expression profiles, with rows representing individual genes
        and columns representing metacells.
}
\description{
This function computes spatial metacells from single-cell spatial transcriptomics.
Spatial metacells represent aggregated expression profiles for spatially proximal cells.
The function utilizes k-nearest neighbor (KNN) weighting to aggregate expression profiles from neighboring cells
and generate metacell expression profiles for each cell type.
}
\examples{
library(data.table)
library(Seurat)
library(SpatialEcoTyper)
library(googledrive)
drive_deauth() # no Google sign-in is required
drive_download(as_id("1CgUOQKrWY_TG61o5aw7J9LZzE20D6NuI"),
                    "HumanMelanomaPatient1_subset_scmeta.tsv", overwrite = TRUE)
drive_download(as_id("1CoQmU3u8MoVC8RbLUvTDQmOuJJ703HHB"),
              "HumanMelanomaPatient1_subset_counts.tsv.gz", overwrite = TRUE)
scdata <- fread("HumanMelanomaPatient1_subset_counts.tsv.gz",
                sep = "\t",header = TRUE, data.table = FALSE)
rownames(scdata) <- scdata[, 1]
scdata <- as.matrix(scdata[, -1])
normdata = NormalizeData(scdata)
scmeta <- read.table("HumanMelanomaPatient1_subset_scmeta.tsv",
                      sep = "\t",header = TRUE, row.names = 1)

# Construct spatial metacells from single-cell spatial data
metacells <- GetSpatialMetacells(normdata = normdata, metadata = scmeta,
                                 X = "X", Y = "Y", CellType = "CellType")
head(metacells)

}
