% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Znorm.R
\name{Znorm}
\alias{Znorm}
\title{Weighted / Unweighted Uni-variance Normalization}
\usage{
Znorm(mat, groups = NULL)
}
\arguments{
\item{mat}{A matrix of gene expression data, where rows represent genes and columns represent cells.}

\item{groups}{A character vector specifying the group labels for each cell.
If provided, weighted univariance normalization will be performed.}
}
\value{
A matrix of Z-score normalized gene expression data, with rows representing genes and columns representing cells.
}
\description{
Z-score normalization scales each feature (gene) across all cells to have a mean of 0 and a standard deviation of 1.
This function allows for optional weighted univariance normalization, where cells are grouped by a categorical variable.
}
\examples{
library(Seurat)
library(data.table)
library(SpatialEcoTyper)
library(googledrive)
drive_deauth() # no Google sign-in is required
drive_download(as_id("1CoQmU3u8MoVC8RbLUvTDQmOuJJ703HHB"),
              "HumanMelanomaPatient1_subset_counts.tsv.gz", overwrite = TRUE)
scdata <- fread("HumanMelanomaPatient1_subset_counts.tsv.gz",
                sep = "\t",header = TRUE, data.table = FALSE)
rownames(scdata) <- scdata[, 1]
scdata <- as.matrix(scdata[, -1])
tmpobj <- CreateSeuratObject(scdata) \%>\%
          SCTransform(clip.range = c(-10, 10), verbose = FALSE)
seurat_version = as.integer(gsub("\\\\..*", "", as.character(packageVersion("SeuratObject"))))
if(seurat_version<5){
  normdata <- GetAssayData(tmpobj, "data")
}else{
  normdata <- tmpobj[["SCT"]]$data
}
# Z-score normalization
znorm_data <- Znorm(normdata)
head(znorm_data[, 1:5])

# Weighted Z-score normalization
drive_download(as_id("12xcZNhpT-xbhcG8kX1QAdTeM9TKeFAUW"),
                     "HumanMelanomaPatient1_subset_scmeta.tsv",
                    overwrite = TRUE, verbose = FALSE)
scmeta <- fread("HumanMelanomaPatient1_subset_scmeta.tsv",
                sep = "\t",header = TRUE, data.table = FALSE)
wtdznorm_data <- Znorm(normdata, groups = scmeta$Region)
head(wtdznorm_data[, 1:5])

}
